+++ tests/tests.py
@@ -1,3 +1,5 @@
+from unittest import skip
+
 from django.template import Context
 from django.test import override_settings, TestCase
 from django.utils import six
@@ -5,7 +7,8 @@ from django.utils import six
 from render_block import render_block_to_string, BlockNotFound, UnsupportedEngine
 
 
-class TestCases(TestCase):
+class TestDjango(TestCase):
+    """Test the Django templating engine."""
     def assertExceptionMessageEquals(self, exception, expected):
         result = exception.message if six.PY2 else exception.args[0]
         self.assertEqual(expected, result)
@@ -38,12 +41,12 @@ class TestCases(TestCase):
 
     def test_include(self):
         """Ensure that an include tag in a block still works."""
-        result = render_block_to_string('test3.html', 'block1')
+        result = render_block_to_string('test3_django.html', 'block1')
         self.assertEqual(result, u'included template')
 
     def test_super(self):
         """Test that block.super works."""
-        result = render_block_to_string('test3.html', 'block2')
+        result = render_block_to_string('test3_django.html', 'block2')
         self.assertEqual(result, u'block2 from test3 - block2 from test1')
 
     def test_subblock(self):
@@ -76,3 +79,69 @@ class TestCases(TestCase):
             render_block_to_string('test1.html', 'noblock')
         self.assertExceptionMessageEquals(exc.exception,
                                           "Can only render blocks from the Django template backend.")
+
+
+@override_settings(
+    TEMPLATES=[{
+        'BACKEND': 'django.template.backends.jinja2.Jinja2',
+        'DIRS': ['tests/templates'],
+        'APP_DIRS': True,
+    },]
+)
+class TestJinja2(TestCase):
+    """Test the Django templating engine."""
+    def assertExceptionMessageEquals(self, exception, expected):
+        result = exception.message if six.PY2 else exception.args[0]
+        self.assertEqual(expected, result)
+
+    def test_block(self):
+        """Test rendering an individual block."""
+        result = render_block_to_string('test1.html', 'block1')
+        self.assertEqual(result, u'block1 from test1')
+
+        # No reason this shouldn't work, but just in case.
+        result = render_block_to_string('test1.html', 'block2')
+        self.assertEqual(result, u'block2 from test1')
+
+    def test_override(self):
+        """This block is overridden in test2."""
+        result = render_block_to_string('test2.html', 'block1')
+        self.assertEqual(result, u'block1 from test2')
+
+    @skip('Not currently supported.')
+    def test_inherit(self):
+        """This block is inherited from test1."""
+        result = render_block_to_string('test2.html', 'block2')
+        self.assertEqual(result, u'block2 from test1')
+
+    def test_no_block(self):
+        """Check if there's no block available an exception is raised."""
+        with self.assertRaises(BlockNotFound) as exc:
+            render_block_to_string('test1.html', 'noblock')
+        self.assertExceptionMessageEquals(exc.exception,
+                                          "block with name 'noblock' does not exist")
+
+    def test_include(self):
+        """Ensure that an include tag in a block still works."""
+        result = render_block_to_string('test3_jinja2.html', 'block1')
+        self.assertEqual(result, u'included template')
+
+    @skip('Not currently supported.')
+    def test_super(self):
+        """Test that super() works."""
+        result = render_block_to_string('test3_jinja2.html', 'block2')
+        self.assertEqual(result, u'block2 from test3 - block2 from test1')
+
+    def test_subblock(self):
+        """Test that a block within a block works."""
+        result = render_block_to_string('test5.html', 'block1')
+        self.assertEqual(result, u'block3 from test5')
+
+        result = render_block_to_string('test5.html', 'block3')
+        self.assertEqual(result, u'block3 from test5')
+
+    def test_context(self):
+        """Test that a context is properly rendered in a template."""
+        data = u'block2 from test5'
+        result = render_block_to_string('test5.html', 'block2', {'foo': data})
+        self.assertEqual(result, data)